Separated Wearables collection client from the regular (now renamed) SceneClient
