+++ unity-renderer/Assets/ABConverter/WearablesCollectionClient.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c1f22c999232c41dc80281e5f6f8c7f4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
